Finished next section of shooter tutorial
This section delt with aiming and crosshairs. We implemented the
functionality to 'look down' the sights (although it's just a zoomed
in perspective from reducing the field of view). We also implemented
dynamic cross hairs that loosen/tighten based on what you're doing.
For example, aiming down sights will tighten the crosshairs, but
shooting will spread them apart.